feat(scripts): script refactoring
